Updated so that constraints are processed while being parsed.
Removed ICspItem and friends.
Removed ExpressionTree and friends.
Updated the parser and the scanner to use IToken directly.
Added Semantics class to handle the processing of constraints, and
    to keep the Parser.y file clean.
Added ConstraintParser class to wrap the parser and to initialize
    Semantics before parsing.
Removed IToken derivatives FunctionToken, OperatorToken, and
    UnknownToken.
Added IToken derivatives BooleanToken, ConstraintToken,
    VariableToken, and IdToken.
Changed NumberToken to store a double instead of an int.
Fixed the way the Scanner handles string literals.
Added a class SpreadSheet, which encapsulates the ConvertAddress
    GetCellValue delegates (which have been removed from the
    CreamCheese class).
Removed the methods AddConstraintsFromETree and
    ContainsConstrainFunction from the CreamCheese class.
Made all classes in CreamCheese explicitly 'internal' or 'public'.